% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_function.R
\name{see_correlation}
\alias{see_correlation}
\title{Fungsi untuk menghitung korelasi dan membuat plot}
\usage{
see_correlation(data, var1, var2, metode = "cor")
}
\arguments{
\item{data}{Data frame yang berisi variabel-variabel yang akan dihitung korelasinya}

\item{var1}{Nama kolom pertama untuk korelasi}

\item{var2}{Nama kolom kedua untuk korelasi}

\item{metode}{Metode yang digunakan untuk menghitung korelasi, bisa 'cor' atau 'cor.test'}
}
\value{
Nilai korelasi atau hasil uji korelasi, tergantung metode yang dipilih, serta plot scatter
}
\description{
Fungsi ini dapat digunakan untuk menghitung korelasi antara dua variabel menggunakan metode `cor()` atau `cor.test()`,
serta membuat plot scatter untuk visualisasi korelasi.
}
\examples{
# Menghitung korelasi dengan metode cor dan membuat plot
see_correlation(mtcars, "mpg", "hp", metode = "cor")

# Menghitung korelasi dengan metode cor.test dan membuat plot
see_correlation(mtcars, "mpg", "hp", metode = "cor.test")
}
